// src/app/accounts/models/ledgerxref.ts
import { GlobalVariables } from '../../core/models/globalvariables';


export class LedgerXref {


    xref_pkid: string;
    xref_jvh_id: string;
    xref_jv_id: string;
    xref_year: number;
    xref_drcr: string;
    xref_acc_id: string;
    xref_dr_jvh_id: string;
    xref_dr_jv_id: string;
    xref_dr_jv_year: number;
    xref_dr_jv_date: string;
    xref_cr_jvh_id: string;
    xref_cr_jv_id: string;
    xref_cr_jv_year: number;
    xref_cr_jv_date: string;
    xref_amt: number;
    xref_adv_amt: number;
  

}

export class XList {
  jvh_id: string;
  XrefList: LedgerXref[] = [];
  _globalvariables: GlobalVariables;
}